refactor(service): split app.py into domain routers (T012)
Extract health, query, job, artifact and atlas routes from the service FastAPI
app into dedicated routers while keeping app.py as composition shell. Health
version fields resolve live from the app module so monkeypatched tests stay
correct. Advances REPOGROUND-LEGACY-RECONCILIATION-V1-T012.
